requirements.txt
attrs==23.2.0
awscli==1.29.45
boto3==1.28.45
botocore==1.31.45
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.4
distlib==0.3.4
docutils==0.16
filelock==3.6.0
idna==3.7
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
mysql-connector-python==8.3.0
numpy==1.25.2
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
pathable==0.4.3
platformdirs==2.5.1
pyasn1==0.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==5.1
referencing==0.31.1
requests==2.31.0
rfc3339-validator==0.1.4
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.6.2
six==1.16.0
urllib3==1.26.16
virtualenv==20.14.0
